+++ src/components/ui/button.tsx
@@ -13,8 +13,8 @@ const button = tv({
 		},
 
 		size: {
-			default: "px-4 py-2.5",
-			sm: "px-3 py-1.5",
+			default: "h-10 px-4",
+			sm: "h-9 px-3",
 		},
 	},